+++ src/app/App.svelte
@@ -80,7 +80,6 @@
     asMedia,
     asFilter,
     asNote,
-    asRelays,
     asRelay,
     asEntity,
   } from "src/app/router"
@@ -147,18 +146,15 @@
   })
   router.register("/notes/:entity", NoteDetail, {
     entity: asNote,
-    relays: asRelays,
   })
   router.register("/notes/:entity/label", LabelCreate, {
     entity: asNote,
   })
   router.register("/notes/:entity/status", PublishInfo, {
     entity: asNote,
-    relays: asRelays,
   })
   router.register("/notes/:entity/thread", ThreadDetail, {
     entity: asNote,
-    relays: asRelays,
   })
   // router.register("/notes/:entity/report", ReportCreate, {
   //   entity: asNote,